Move Psbt to new v0 module
In preparation for adding PSBTv2, move the current logic to a new `v0`
submodule. Add a type alias from the crate root.
